Move/rename firmware extractor tools.
Since sigrok-util will be released as tarball with the usual build
system at some point, and the tools will be installed in /usr/bin by
most distros, the tools need to have unique names as to not conflict
with other things in /usr/bin. We use a common 'sigrok-fwextract-' prefix
for all current and future firmware extractor tools to keep the
names consistent and unique.

Drop the .py suffix from the executables, since the implementation
language doesn't matter and is generally not encoded in the filename
in things installed in /usr/bin.

Later on, every tool will also get a manpage, of course.