sigrok-cross-mingw: i686-pc-mingw32.static -> i686-w64-mingw32.static.
We're switching from MinGW (i686-pc-mingw32.static in MXE)...

  http://mingw.org/

...to MinGW-w64 (i686-w64-mingw32.static in MXE):

  http://mingw-w64.sourceforge.net/

The latter is a fork which nowadays seems to be more active and also to
have a more complete set of win32 API headers and libs.

History:
http://sourceforge.net/p/mingw-w64/wiki2/History/

Among other things this is required for certain API calls now used in
recent libserialport versions (probably will be required for various
other sigrok-related projects later as well).