new-driver: fix for libsigrok since dd5c48a6
This patch in libsigrok removes the drivers.c completely, so new-driver
should not try to modify it any more. It has been replaced by a register
mechanism to add driver pointer in a new special section.
So drv-api.c must be modified to include a call to SR_REGISTER_DEV_DRIVER