sigrok-cross-android: Set "android:versionName" to a timestamp.
Use the current date-/timestamp when building as the versionName
Android field.

This "version" basically means "git HEAD of everything as of the
build time of the APK", which includes every piece of software that
the sigrok-cross-android script grabs and builds, as well as the
pre-installed Android NDK + SDK components.