sigrok-native-msys2: Have libsigrok Python bindings use Python 3.
We're installing Python 3 versions of setuptools and numpy using this
script's "prepare" option, so force the build to use those (otherwise
it'll search for the Python 2 versions and won't find them).

Note: The Python bindings build still remains off (--disable-python)
since there are issues there that need to be investigated.