sigrok-native-appimage: Bump to use Ubuntu 16.04 as base OS.
Until recently, the sigrok AppImages were built on Ubuntu 14.04
Trusty Tahr (originally released 2014, Canonical support ended 04/2019).

We're now building on Ubuntu 16.04 Xenial Xerus (originally released
2016, Canonical support will end 04/2021).

https://wiki.ubuntu.com/Releases

This release is still old enough so that the generated AppImages should
run on the majority of modern Linux distros, but also recent enough so
we somewhat recent versions of all libraries. In the case of Qt, we get
a recent enough version to be able to build PulseView (which currently
needs Qt >= 5.3) at all: Qt 5.5.1.