firmware: kingst-la: remove README
In general, the idea of having a README is nice so people can get easily
information when browsing the repo in github. However, it is
error-prone because it requiers a manual step to have the README and
manpage always in sync. Even if we agree to have READMEs, they should be
done for all firmware tools and, thus, the generation would need a
different approach. Let's keep things consistent for now.

Signed-off-by: Wolfram Sang <wsa@kernel.org>